Release branch only commit: Frame barrier fix
Ensure the frame barrier prevents reordering of stack pointer changes

It is possible for a load/store accessing the stack via a copy of the
stack pointer to be moved across the epilogue meaning that it accesses
stack that is no longer allocated. This leads to a situation where the
code is unsafe in the event of an interrupt where the same stack is
used for interrupt handling.

gcc/
	* config/mips/mips.c (mips_frame_barrier): Upgrade to a full
	blockage.

(cherry picked from commit 9a21c4952364d368efdb955a6d62ba64e674a84d)