GT-110: Adjust the m constraint to be usable for all load/store insns
End user expect that the 'm' constraint will construct a reference to
memory that can be used with any load or store instructions. Owing to
the range of addressing modes available in nanoMIPS and the need to
match the mode with the mnemonic it is not possible to support all
addressing modes with one constraint.  Instead restrict the 'm'
constraint in inline ASM to use just a base register with no offset
or index.

gcc/
	* doc/md.texi (constraints): Add nanoMIPS constraint definitions
	mirroring the MIPS versions.
	* config/mips/mips.c (mips_md_asm_adjust): New static function.
	[NANOMIPS_SUPPORT] (TARGET_MD_ASM_ADJUST): New macro.

gcc/testsuite/
	* gcc.target/nanomips/constraint-m.c: New file.

(cherry picked from commit 1f751b82998cb5755f88d317e838c3284bfe9f53)