Fix PIC addressing mode for string constants
The string constant optimisation to use LI48 to calculate the address
was not suitable for PIC code. Disable it and rely on the linker to
perform the optimisation for PIC code as it will be a simple got_disp
to translate to an LI48/LAPC48.

gcc/
	* config/mips/mips.c (mips_string_constant_p): Disable the
	optimisation for PIC code.

(cherry picked from commit 5954ed4e45ebce15fb1de97351adddbb9f6e9258)